Separate out objects and begin work on testing
refs #31994
