+++ python/moosepy/perfgraph/__init__.py
@@ -1,0 +1,14 @@
+# This file is part of the MOOSE framework
+# https://mooseframework.inl.gov
+#
+# All rights reserved, see COPYRIGHT for full restrictions
+# https://github.com/idaholab/moose/blob/master/COPYRIGHT
+#
+# Licensed under LGPL 2.1, please see LICENSE for details
+# https://www.gnu.org/licenses/lgpl-2.1.html
+
+from .perfgraph import PerfGraph
+from .perfgraphnode import PerfGraphNode
+from .perfgraphsection import PerfGraphSection
+
+__all__ = ["PerfGraphNode", "PerfGraphSection", "PerfGraph"]
